+++ CHANGELOG.md
@@ -20,3 +20,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spec now v0_4_1
 - Included CLAUDE.md
 - Phase 0-2 complete
+- Phase 3 complete